Replace glib with Objectively + SDL3 throughout
Removes glib-2.0 dependency across all modules, replacing:
- GList/GSList → Objectively List
- GArray/GPtrArray → Objectively Vector
- GHashTable → Objectively HashTable
- g_str_has_prefix/suffix → strncmp/strcmp
- g_ascii_strcasecmp → SDL_strcasecmp
- g_strsplit → SDL_strtok_r + Vector
- g_uuid_string_random → SDL_rand-based RFC 4122 v4 UUID
- g_str_hash → djb2 hash
- g_strstrip → manual trim
- gulong → size_t

Also fixes:
- md5_finalize bug: total_bits now computed before padding update
- Atlas_NodeComparator: correctly dereferences atlas_node_t** pointers
- All Makefile.am files: @GLIB_CFLAGS@/@GLIB_LIBS@ → @OBJECTIVELY_CFLAGS@/@OBJECTIVELY_LIBS@
- configure.ac: GLIB pkg-config check removed

All 19 tests pass.

+++ src/client/cl_screen.c
@@ -319,7 +319,7 @@ static void Cl_DrawSoundStats(void) {
 /**
  * @brief Formats a stat counter string showing the current value and min/max over the sample window.
  */
-static void Cl_DrawSampleCounter(char *buffer, gulong buffer_length, const char *title, uint16_t *samples) {
+static void Cl_DrawSampleCounter(char *buffer, size_t buffer_length, const char *title, uint16_t *samples) {
   uint16_t min, max;
 
   if (!cl.sample_count) {
@@ -347,7 +347,7 @@ static void Cl_DrawSampleCounter(char *buffer, gulong buffer_length, const char 
 /**
  * @brief Formats a frame time counter string showing the current value and min/max over the sample window.
  */
-static void Cl_DrawFrameTimeSampleCounter(char *buffer, gulong buffer_length, const char *title, uint8_t *samples) {
+static void Cl_DrawFrameTimeSampleCounter(char *buffer, size_t buffer_length, const char *title, uint8_t *samples) {
   uint16_t min, max;
 
   if (!cl.sample_count) {
